chore(jira): rename scripts from .mjs to .js, switch to named js-yaml imports
- Renamed api/apply/fetch/process from .mjs to .js to match repo convention
  (package.json has "type": "module" so .js files are already ESM)
- Switched from default import to named imports for js-yaml v5 compatibility:
  `import { load as yamlLoad, dump as yamlDump } from "js-yaml"`
- Added js-yaml as an explicit root devDependency (was previously a hoisted
  transitive dep from read-yaml-file@js-yaml@3.x)
- Updated workflow and CLAUDE.md script references from .mjs to .js
