+++ package.json
@@ -77,6 +77,7 @@
     "globals": "^17.6.0",
     "happy-dom": "^20.9.0",
     "husky": "^9.1.7",
+    "js-yaml": "^5.1.0",
     "minimatch": "^10.2.5",
     "msw": "^2.14.3",
     "playwright": "^1.60.0",
