+++ scripts/jira/process.spec.js
@@ -14,7 +14,7 @@ import { describe, it, expect, vi } from "vitest";
 import { writeFileSync, existsSync, unlinkSync } from "fs";
 import { join } from "path";
 import { tmpdir } from "os";
-import { processFile } from "./process.mjs";
+import { processFile } from "./process.js";
 
 function mockApi(overrides = {}) {
   return {
